Opportunity component, backend and frontend

+++ resources/js/components/views/bookings/index.vue
@@ -311,7 +311,7 @@
                 drivers: {},
                 model: {},
                 columns: {},
-                source: '/api/bookings',
+                source: '/api/v1/bookings',
                 query: {
                     page: 1,
                     column: 'id',
@@ -360,6 +360,9 @@
                     });
             },
             deleteBooking(id) {
+                if(confirm('are you sure?'))
+
+                // Send request to the server
                 axios.delete( '/api/v1/bookings/'+id)
                     .then(function (response) {
                         window.location.reload();
@@ -391,7 +394,7 @@
             fetchIndexData() {
                 var vm = this;
 
-                const url = 'http://127.0.0.1:8000/api/bookings?column=' + this.query.column + '&direction=' + this.query.direction + '&page=' + this.query.page + '&per_page=' + this.query.per_page + '&search_column=' + this.query.search_column + '&search_operator=' + this.query.search_operator + '&search_input=' + this.query.search_input;
+                const url = 'http://127.0.0.1:8000/api/v1/bookings?column=' + this.query.column + '&direction=' + this.query.direction + '&page=' + this.query.page + '&per_page=' + this.query.per_page + '&search_column=' + this.query.search_column + '&search_operator=' + this.query.search_operator + '&search_input=' + this.query.search_input;
 
                 axios.get(url)
                     .then(function(response) {
